backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
sentence-transformers>=2.6.1
scikit-learn>=1.4.2
nltk>=3.8.1
pymupdf>=1.24.1
python-docx>=1.1.0
pandas>=2.2.2
numpy>=1.26.4
reportlab>=4.2.2
pydantic>=2.7.0
transformers>=4.44.0
torch>=2.2.0
